pyproject.toml
[tool.poetry]
name = "demo-ai-assistant"
version = "0.1.0"
description = ""
authors = ["Denis Voskvitsov <denis.voskvitsov@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
llama-cpp-python = "^0.2.79"
huggingface-hub = "^0.23.4"
requests = "^2.32.3"
python-decouple = "^3.8"
langchain-community = "^0.2.6"
langchain-mistralai = "^0.1.9"
langchain-experimental = "^0.0.62"
griptape = {extras = ["all"], version = "^0.27.2"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
